requirements.txt
# ============================
# 🧠 Emotion Recognition Therapy App
# Optimized for Render.com Deployment
# ============================

# --- Web Framework & Utilities ---
Flask==3.0.3
Flask-Cors==4.0.0
Werkzeug==3.0.3
gunicorn==22.0.0

# --- Machine Learning / Deep Learning ---
# Using TensorFlow 2.16.x for better stability on Render
tensorflow==2.16.2
keras==3.3.3
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0

# --- Data Management & Analysis ---
pandas==2.0.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib
tqdm

# --- Firebase / Firestore Integration ---
firebase-admin==6.4.0
google-cloud-firestore==2.16.0
google-api-core==2.19.1

# --- AI Integration ---
google-generativeai==0.8.5
requests==2.32.3
urllib3==2.2.2

# --- Security & Authentication ---
bcrypt==4.1.3
itsdangerous==2.2.0

# --- Utilities ---
python-dotenv==1.0.0
python-dateutil==2.9.0.post0

# Note: matplotlib, tqdm, pytest removed for faster deployment